chore: move components to corresponding folders
