+++ ClientApp/src/app/modules/home/Components/chat/chat.component.ts
@@ -2,11 +2,7 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-chat',
-  standalone: true,
-  imports: [],
   templateUrl: './chat.component.html',
   styleUrl: './chat.component.css'
 })
-export class ChatComponent {
-
-}
+export class ChatComponent { }
