+++ ClientApp/src/app/modules/home/Components/dropdown-menu/dropdown-menu.component.ts
@@ -3,8 +3,6 @@ import { Component, ElementRef, HostListener, Input, Renderer2, ViewContainerRef
 
 @Component({
   selector: 'app-dropdown-menu',
-  standalone: true,
-  imports: [ CommonModule ],
   templateUrl: './dropdown-menu.component.html',
   styleUrl: './dropdown-menu.component.css'
 })
