+++ ClientApp/src/app/modules/home/Components/round-button/round-button.component.ts
@@ -2,8 +2,6 @@ import { Component, Input } from '@angular/core';
 
 @Component({
   selector: 'app-round-button',
-  standalone: true,
-  imports: [],
   templateUrl: './round-button.component.html',
   styleUrl: './round-button.component.css'
 })
