+++ ClientApp/src/app/modules/home/Components/search-bar/search-bar.component.html
