+++ ClientApp/src/app/modules/home/Components/search-bar/search-bar.component.ts
@@ -2,8 +2,6 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-search-bar',
-  standalone: true,
-  imports: [],
   templateUrl: './search-bar.component.html',
   styleUrl: './search-bar.component.css'
 })
