+++ ClientApp/src/app/modules/home/Pages/friends-sidebar/friends-sidebar.component.ts
